% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{shinyresearch_packages}
\alias{shinyresearch_packages}
\title{List all packages imported by shinyresearch}
\usage{
shinyresearch_packages()
}
\description{
List all packages imported by shinyresearch
}
\examples{
shinyresearch_packages()
}
